+++ src/Sophia_Script_for_Windows_11/Sophia.ps1
@@ -2,8 +2,8 @@
 	.SYNOPSIS
 	Default preset file for "Sophia Script for Windows 11"
 
-	Version: v6.3.1
-	Date: 06.02.2023
+	Version: v6.3.2
+	Date: 11.02.2023
 
 	Copyright (c) 2014—2023 farag
 	Copyright (c) 2019—2023 farag & Inestic
@@ -26,8 +26,8 @@
 
 	.NOTES
 	Supported Windows 11 versions
-	Versions: 21H2/22H2
-	Builds: 22000.1335+, 22621.963+
+	Versions: 22H2
+	Builds: 22621.963+
 	Editions: Home/Pro/Enterprise
 
 	.NOTES
@@ -69,7 +69,7 @@ param
 
 Clear-Host
 
-$Host.UI.RawUI.WindowTitle = "Sophia Script for Windows 11 v6.3.1 | Made with $([char]::ConvertFromUtf32(0x1F497)) of Windows | $([char]0x00A9) farag & Inestic, 2014$([char]0x2013)2023"
+$Host.UI.RawUI.WindowTitle = "Sophia Script for Windows 11 v6.3.2 | Made with $([char]::ConvertFromUtf32(0x1F497)) of Windows | $([char]0x00A9) farag & Inestic, 2014$([char]0x2013)2023"
 
 Remove-Module -Name Sophia -Force -ErrorAction Ignore
 Import-Module -Name $PSScriptRoot\Manifest\Sophia.psd1 -PassThru -Force
@@ -90,7 +90,7 @@ Import-LocalizedData -BindingVariable Global:Localization -BaseDirectory $PSScri
 #>
 if ($Functions)
 {
-	#Invoke-Command -ScriptBlock {Checks}
+	Invoke-Command -ScriptBlock {Checks}
 
 	foreach ($Function in $Functions)
 	{